fix: align Ctrl+Click navigation flow with F12 for smoother and consistent definition jumps
